Refactor: 커뮤니티 더미 데이터 파일 분리
