+++ app/dummy/community.ts
@@ -1,0 +1,255 @@
+import { CommunityPostType } from '@/types';
+
+export const CommunityPostDummy: CommunityPostType[] = [
+  {
+    title: '혹시 이번에 웅진씽크빅 합격 연락 받은사람?ㅜㅠ',
+    content:
+      '내가 이것저것 대외활동도 하고 면접도 나쁘지 않았는데, 아직 결과 안온듯,,ㅜㅠ 문자 받은사람있으신가요?',
+    tags: ['회사', '취업', '인턴', '신입', '채용', '면접'],
+    category: '질의응답',
+    createdAt: '2024-01-02 18:30',
+    likes: 224,
+    comments: [
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+    ],
+    views: 2680,
+  },
+  {
+    title: 'UXUI 인턴형 프로그램 1기 후기',
+    content:
+      '이번에 인턴형 프로그램 1기에 참여한 무순죽순채순입니다. 10월부터 사전 직무교육 후에 팀 프로젝트를 진행했습니다.',
+    tags: ['UX', '디자인', '인턴형프로그램', '후기', '취업', 'UI'],
+    category: '스팩 후기',
+    createdAt: '2024-01-02 18:30',
+    likes: 12,
+    comments: [
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+    ],
+    views: 1892,
+  },
+  {
+    title: '대학교에서 졸업준비위원회 한거 포폴에 넣어도 되나?',
+    content:
+      '내가 대학교에서 졸준위했는데 다같이 작업하긴 했는데 거의 내 디자인임. 그래서 포폴에 넣을건데 상관없나?',
+    tags: ['디자인', '포트폴리오', '졸업전시'],
+    category: '자유게시판',
+    createdAt: '2024-01-09 18:20',
+    likes: 23,
+    comments: [
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+    ],
+    views: 1720,
+    thumbnail:
+      'https://images.unsplash.com/photo-1609921212029-bb5a28e60960?q=80&w=2052&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
+  },
+  {
+    title: '서브프로젝트 성공적인 후기',
+    content:
+      '드디어 성공적인 서브 프로젝트를 진행했다. 솔직히 여기저기서 힘들다는 이야기 많았지만, 이정도 일 줄은,, 그래도 작업물 보면 뿌듯함.',
+    tags: ['서브프로젝트', '디자인', '후기'],
+    category: '자유게시판',
+    createdAt: '2024-01-02 18:30',
+    likes: 46,
+    comments: [
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+    ],
+    views: 1588,
+  },
+  {
+    title:
+      '이번에 개발자로 취업하고 싶어서 스팩에서 개발자 대외활동 신청할건데 같이할 사람',
+    content:
+      '나는 어느정도 기초만 할 줄 아는데, 더 배우고 싶은 마음도 있고, 새로운 사람들도 많이 만나보고 싶음...',
+    tags: ['개발자', '부트캠프', '대외활동', '취업'],
+    category: '자유게시판',
+    createdAt: '2024-01-02 18:30',
+    likes: 11,
+    comments: [
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+    ],
+    views: 1243,
+  },
+  {
+    title: 'UXUI 인턴형 프로그램 2기 후기',
+    content:
+      '내가 1기 놓치고 2기가 생각보다 금방 시작해서 지원했는데, 합격해서 진짜 열심히함. 1기보다 직무교육도 짧아서 걱정했는데, 나쁘지 않았음...',
+    tags: ['UX', '디자인', '인턴형프로그램', '후기', '취업', 'UI'],
+    category: '자유게시판',
+    createdAt: '2024-01-06 18:30',
+    likes: 187,
+    comments: [
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+    ],
+    views: 1176,
+    thumbnail:
+      'https://images.unsplash.com/photo-1609921212029-bb5a28e60960?q=80&w=2052&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
+  },
+  {
+    title: '회사 이직..그거 어떻게 하는건데..',
+    content:
+      '이직할건데 일한지 이제 1년 4개월 넘음, 지금하면 좀 그런가,, 다들 언제 이직했어ㅜㅠㅜㅠㅜㅠ',
+    tags: ['회사', '이직', '퇴사'],
+    category: '자유게시판',
+    createdAt: '2024-01-06 18:30',
+    likes: 20,
+    comments: [
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+    ],
+    views: 1023,
+  },
+  {
+    title: '개발자 부트캠프 후기',
+    content:
+      '개발자로 일하다가 쉬어갈겸, 부트캠프해서 실력 좀 키우고 왔는데, 역시 스팩 부트캠프는 진짜 믿고 하는 듯, 수준도 좋고 팀원들도 잘 만났음...',
+    tags: ['개발자', '부트캠프', '후기', '취업'],
+    category: '자유게시판',
+    createdAt: '2024-01-09 14:30',
+    likes: 12,
+    comments: [
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+    ],
+    views: 875,
+    thumbnail:
+      'https://images.unsplash.com/photo-1609921212029-bb5a28e60960?q=80&w=2052&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
+  },
+  {
+    title: '대학생 3학년인데 대외활동 2개함,,,',
+    content:
+      '미대생이라 과제가 너무 많은데 이번에 3학년 올라가는데 대외활동 두개함,,이게 맞는건가,,,',
+    tags: ['대학생', '디자인', '대외활동'],
+    category: '자유게시판',
+    createdAt: '2024-01-09 14:30',
+    likes: 21,
+    comments: [
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+    ],
+    views: 731,
+  },
+  {
+    title: '대기업을 퇴사한 이유',
+    content:
+      '저는 누구나 들으면 아는 대기업에서 디자이너로 일하고 있었습니다. 다른 사람들 모두 퇴사하는 이유가 있는데 저도 어린나이에도 불구하고 퇴사라...',
+    tags: ['UX', '디자인', '회사', '디자이너', 'UI'],
+    category: '자유게시판',
+    createdAt: '2024-01-09 14:30',
+    likes: 590,
+    comments: [
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+    ],
+    views: 650,
+  },
+  {
+    title: '서브 프로젝트가 경력으로 들어가나요?',
+    content:
+      '제가 이력서나 포폴에 서브프로젝트를 적으려하는데 이걸 경력으로 쳐주나요? 잘 모르겠어요ㅜㅠ',
+    tags: ['경력', '포트폴리오', '디자이너', '취업'],
+    category: '자유게시판',
+    createdAt: '2024-01-08 14:30',
+    likes: 12,
+    comments: [
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+    ],
+    views: 632,
+    thumbnail:
+      'https://images.unsplash.com/photo-1609921212029-bb5a28e60960?q=80&w=2052&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
+  },
+  {
+    title: 'UXUI 인턴형 프로그램 1기 후기',
+    content:
+      '이번에 인턴형 프로그램 1기에 참여한 무순죽순채순입니다. 10월부터 사전 직무교육 후에 팀 프로젝트를 진행했습니다.',
+    tags: ['UX', '디자인', '인턴형프로그램', '후기', '취업', 'UI'],
+    category: '스팩 후기',
+    createdAt: '2024-01-02 18:30',
+    likes: 10,
+    comments: [
+      { name: '', content: '', createdAt: '', likes: 0, comments: [] },
+    ],
+    views: 500,
+  },
+];
